drop separate cart totals resolver
