+++ libs/domain/cart/src/services/cart.service.ts
@@ -6,14 +6,14 @@ import {
   CartEntry,
   CartEntryQualifier,
   CartQualifier,
-  CartTotals,
+  NormalizedTotals,
   UpdateCartEntryQualifier,
 } from '../models';
 
 export interface CartService {
   getCart(data?: CartQualifier): Observable<Cart | undefined>;
   getCartState(data?: CartQualifier): Observable<QueryState<Cart>>;
-  getTotals(data?: CartQualifier): Observable<CartTotals | null>;
+  getTotals(data?: CartQualifier): Observable<NormalizedTotals | null>;
   getEntries(data?: CartQualifier): Observable<CartEntry[]>;
   addEntry(data: AddCartEntryQualifier): Observable<unknown>;
   updateEntry(data: UpdateCartEntryQualifier): Observable<unknown>;
