+++ libs/domain/cart/src/services/totals/cart-totals.providers.ts
@@ -1,0 +1,8 @@
+import { Provider } from '@spryker-oryx/di';
+import { CartService } from '../cart.service';
+import { TotalsResolver } from './totals.service';
+
+export const CartTotalsProvider: Provider = {
+  provide: `${TotalsResolver}CART`,
+  useExisting: CartService,
+};
