+++ libs/domain/cart/src/services/totals/index.ts
@@ -1,3 +1,3 @@
-export * from './cart-totals.resolver';
+export * from './cart-totals.providers';
 export * from './default-totals.service';
 export * from './totals.service';
